File reading finished; added hardness

+++ src/room.h
@@ -22,7 +22,7 @@ typedef struct
 } room;
 
 bool room_overlaps(room* first, room* second);
-bool room_contains_room(int x, int y, room* place);
+bool room_contains_room(uint8_t x, uint8_t y, room* place);
 int room_lazy_distance(room* first, room* second);
 
 #endif
